build: fix compiler warning Bar.volume is long long

+++ src/tws_meta.h
@@ -300,7 +300,7 @@ struct RowHist
 	double high;
 	double low;
 	double close;
-	int volume;
+	long long volume;
 	int count;
 	double WAP;
 	bool hasGaps;
